* ARTTARIFLIGNE.cs: working on buy
* TIERS.cs:
* PIECEACHAT.cs:
* PIECEVENTE.cs:
* DataServices.cs:
* RestServices.cs:
* REGLEMENTECHEANCE.cs:
* PIECEVENTEECHEANCE.cs:
* PIECEACHATECHEANCE.cs:
* BuyFRentPageModel.cs:
* BuyFAentPageModel.cs:
* BuyBCLigPageModel.cs:
* BuyBCEntPageModel.cs:
* SellBREntPageModel.cs:
* SellFREntPageModel.cs:
* SellFVEntPageModel.cs:
* SellLignePageModel.cs:
* SellBLEntPageModel.cs:
* BuyDetailsPageModel.cs:
* SellEntetePageModel.cs:
* StockLignePageModel.cs:
* SellDetailsPageModel.cs:
* BuyBRecentetPageModel.cs:
* BuyBRetEntetPageModel.cs:

+++ PFE/PFE/Models/ARTTARIFLIGNE.cs
@@ -7,7 +7,7 @@ namespace PFE.Models
         public int TRFID { get; set; }
         public int ATFLIGNE { get; set; }
         public int ATFQTE { get; set; }
-        public int ATFPRIX { get; set; }
+        public double ATFPRIX { get; set; }
         public string ATFFORMULE { get; set; }
     }
 }
