+++ PFE/PFE/Models/PIECEACHATECHEANCE.cs
@@ -7,7 +7,7 @@ namespace PFE.Models
         public int? PEAID { get; set; }
         public DateTime? PEADATE { get; set; }
         public float? PEAMONTANT { get; set; }
-        public int? PEATAUX { get; set; }
+        public double? PEATAUX { get; set; }
         public int? RGTID { get; set; }
         public string PEAISREGLE { get; set; }
         public int? PEAMNTREGLE { get; set; }
