+++ PFE/PFE/Models/PIECEVENTEECHEANCE.cs
@@ -6,8 +6,8 @@ namespace PFE.Models
         public int? PCVID { get; set; }
         public int? PEVID { get; set; }
         public DateTime? PEVDATE { get; set; }
-        public int? PEVMONTANT { get; set; }
-        public int? PEVTAUX { get; set; }
+        public double? PEVMONTANT { get; set; }
+        public double? PEVTAUX { get; set; }
         public int? RGTID { get; set; }
         public string PEVISREGLE { get; set; }
         public int? PEVMNTREGLE { get; set; }
